Freeplane1.7.5 Format: Fix fork background colour
In script for Freeplane 1.3 set background colour for nodes with fork
style to null, which is transparent, so they would pick up the colour
of the cloud or the map behind them.  Now if default style is applied
and background colour is set to null the node will pick up the colour
defined in the style.

Rather than attempting to modify the styles we simple set the
node background colour to be the same as the cloud or map behind it.